+++ .github/workflows/forkprbuildpack.yml
@@ -45,7 +45,7 @@ jobs:
       mmc: ${{ steps.artifactNames.outputs.mmc }}
     steps:
       - name: Checkout Ref
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
         with:
           fetch-depth: 0
           ref: ${{ github.event.pull_request.head.sha }}
@@ -54,9 +54,9 @@ jobs:
       # Don't use cache to prevent cache poisoning
 
       - name: Setup NodeJS v20
-        uses: actions/setup-node@v4
+        uses: actions/setup-node@v6
         with:
-          node-version: 20
+          node-version: 22
           check-latest: true
 
       - name: Setup NPM Packages
@@ -80,7 +80,7 @@ jobs:
     needs: setup
     steps:
       - name: Checkout Ref
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
         with:
           fetch-depth: 0
           ref: ${{ github.event.pull_request.head.sha }}
@@ -89,9 +89,9 @@ jobs:
       # Don't use cache to prevent cache poisoning
 
       - name: Setup NodeJS v20
-        uses: actions/setup-node@v4
+        uses: actions/setup-node@v6
         with:
-          node-version: 20
+          node-version: 22
           check-latest: true
 
       - name: Setup NPM Packages
@@ -105,7 +105,7 @@ jobs:
           CFCORE_API_TOKEN: ${{ secrets.CFCORE_API_TOKEN }}
 
       - name: Upload Client Zip
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: ${{ needs.setup.outputs.client }}
           path: ./build/client/**/*
@@ -118,7 +118,7 @@ jobs:
     needs: setup
     steps:
       - name: Checkout Ref
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
         with:
           fetch-depth: 0
           ref: ${{ github.event.pull_request.head.sha }}
@@ -127,9 +127,9 @@ jobs:
       # Don't use cache to prevent cache poisoning
 
       - name: Setup NodeJS v20
-        uses: actions/setup-node@v4
+        uses: actions/setup-node@v6
         with:
-          node-version: 20
+          node-version: 22
           check-latest: true
 
       - name: Setup NPM Packages
@@ -143,7 +143,7 @@ jobs:
           CFCORE_API_TOKEN: ${{ secrets.CFCORE_API_TOKEN }}
 
       - name: Upload Server Zip
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: ${{ needs.setup.outputs.server }}
           path: ./build/server/**/*
@@ -156,7 +156,7 @@ jobs:
     needs: setup
     steps:
       - name: Checkout Ref
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
         with:
           fetch-depth: 0
           ref: ${{ github.event.pull_request.head.sha }}
@@ -165,9 +165,9 @@ jobs:
       # Don't use cache to prevent cache poisoning
 
       - name: Setup NodeJS v20
-        uses: actions/setup-node@v4
+        uses: actions/setup-node@v6
         with:
-          node-version: 20
+          node-version: 22
           check-latest: true
 
       - name: Setup NPM Packages
@@ -181,7 +181,7 @@ jobs:
           CFCORE_API_TOKEN: ${{ secrets.CFCORE_API_TOKEN }}
 
       - name: Upload Lang Zip
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v5
         with:
           name: ${{ needs.setup.outputs.lang }}
           path: ./build/lang/**/*
